backend/requirements/development.txt
-r base.txt
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.1
Faker==26.0.0
freezegun==1.5.1
responses==0.25.3
django-debug-toolbar==4.4.2
ipython==8.26.0
